requirements.txt
Flask==3.0.0
google-generativeai==0.4.0